+++ backend/src/mappening/utils/locations.py
@@ -1,12 +1,12 @@
 from mappening.utils.database import *
+from mappening.utils import tokenize
+import location_utils
 
 import requests
 import re
 import json
 import os
 from operator import itemgetter
-import location_utils
-from mappening.utils import tokenize
 
 # Must be on same level as app.py
 data = json.load(open('sampleData.json'))
